src/requirements.txt
absl-py==0.7.1
albumentations==0.4.2
alembic==1.0.11
amqp==2.5.1
aniso8601==7.0.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.0.5
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-crispy-forms==1.8.0
django-debug-permissions==1.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==19.9.0
idna==2.10
imageio==2.9.0
imgaug==0.2.6
Jinja2==2.10.1
kiwisolver==1.2.0
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
networkx==2.4
numpy==1.19.0
oauthlib==3.1.0
opencv-python-headless==4.3.0.36
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.4.0
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.2.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
scikit-image==0.17.2
scipy==1.5.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.18
sqlparse==0.3.1
tifffile==2020.7.4
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
